speedometer_app1/requirements.txt
fastapi==0.95.0
uvicorn==0.21.1
paho-mqtt==2.1.0